+++ Quicksilver/Code-QuickStepCore/QSLibrarian.m
@@ -17,7 +17,7 @@
 
 float QSMinScore = 0.333333;
 
-static int presetSort(id item1, id item2, QSLibrarian *librarian) {
+static NSInteger presetSort(id item1, id item2, QSLibrarian *librarian) {
 	return [[item1 name] caseInsensitiveCompare:[item2 name]];
 }
 
@@ -153,7 +153,7 @@ static float searchSpeed = 0.0;
 		}
 		children = [parent getChildren];
 		[children addObject:entry];
-		[children sortUsingFunction:(int (*)(id, id, void *))presetSort context:(void *)self];
+		[children sortUsingFunction:(NSInteger (*)(id, id, void *))presetSort context:(void *)self];
 		if (scan) [entry scanForced:YES];
 	}
 	//[catalogChildren replaceObjectsInRange:NSMakeRange(0, 0) withObjectsFromArray:newPresets];
