+++ Quicksilver/Code-QuickStepInterface/QSDelegatingTableColumn.m
@@ -14,7 +14,7 @@
 @end
 
 @implementation QSDelegatingTableColumn
-- (id)dataCellForRow:(int)row {
+- (id)dataCellForRow:(NSInteger)row {
 	id delegate = [[self tableView] delegate];
 	NSCell *cell = nil;
 	if ([[self tableView] isKindOfClass:[NSOutlineView class]]) {
