+++ Quicksilver/Code-QuickStepInterface/QSLargeTypeDisplay.m
@@ -113,7 +113,7 @@ void QSShowLargeType(NSString *number) {
 
 
 @implementation QSVanishingWindow
-- (id)initWithContentRect:(NSRect)contentRect styleMask:(unsigned int)aStyle backing:(NSBackingStoreType)bufferingType defer:(BOOL)flag {
+- (id)initWithContentRect:(NSRect)contentRect styleMask:(NSUInteger)aStyle backing:(NSBackingStoreType)bufferingType defer:(BOOL)flag {
 	if (self = [super initWithContentRect:contentRect styleMask:aStyle backing:bufferingType defer:flag]) {
 		[self setReleasedWhenClosed:YES];
 	}
